refactor: Improve backend and frontend architecture
This commit introduces a significant refactoring of the application to improve its structure and maintainability.

Backend:
- The backend has been restructured to follow a modular, MVC-like pattern.
- All business logic has been moved from `index.js` into separate controller files within the `controllers` directory.
- All route definitions have been moved to separate route files within the `routes` directory.
- Authentication middleware has been extracted into its own file in the `middleware` directory.
- The main `index.js` file is now much cleaner and primarily responsible for server setup and middleware configuration.

Frontend:
- The monolithic `Monitor.jsx` component has been broken down into smaller, more focused components (`ClientView`, `ClientJobsView`, `OperationsView`, etc.).
- Data fetching and state management logic have been extracted into custom hooks (`useClients`, `useJobs`) to improve reusability and separation of concerns.
- Utility functions have been moved to a dedicated `utils` directory.
- The overall frontend codebase is now more modular, easier to understand, and follows React best practices.
